refactor sub-menus
