--- src/containers/hover-menu/constants.tsx
@@ -1,17 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-
-import CHART_SVG from 'svgs/sidebar/chart.svg?sprite';
-import RELOAD_SVG from 'svgs/sidebar/reload.svg?sprite';
-import STAR_SVG from 'svgs/sidebar/star.svg?sprite';
-import SUN_SVG from 'svgs/sidebar/sun.svg?sprite';
-
-export const WIDGET_OPTIONS = [
-  { id: 'summary', label: 'Summary', icon: STAR_SVG },
-  {
-    id: 'restoration_and_conservation',
-    label: 'Restoration & Conservation',
-    icon: RELOAD_SVG,
-  },
-  { id: 'climate_and_policy', label: 'Climate & Policy', icon: SUN_SVG },
-  { id: 'ecosystem_services', label: 'Ecosystem Services', icon: CHART_SVG },
-];
