--- src/containers/sidebar/menu/constants.tsx
@@ -1,32 +1,0 @@
-export const EXT_MENU_OPTIONS = [
-  {
-    id: 'about',
-    label: 'About',
-    href: 'https://www.mangrovealliance.org/about-us',
-  },
-  {
-    id: 'mangroves',
-    label: 'Mangroves',
-    href: 'https://www.mangrovealliance.org/mangrove-forests',
-  },
-  {
-    id: 'initiatives',
-    label: 'Initiatives',
-    href: 'https://www.mangrovealliance.org/initiatives',
-  },
-  {
-    id: 'news',
-    label: 'News',
-    href: 'https://www.mangrovealliance.org/news',
-  },
-  {
-    id: 'resources',
-    label: 'Resources',
-    href: 'https://www.mangrovealliance.org/tools-and-resources',
-  },
-  {
-    id: 'contact',
-    label: 'Contact',
-    href: 'https://www.mangrovealliance.org/contact',
-  },
-];
